Rearrange steps
